+++ src/analyzer/expr/assignment/array_assignment_analyzer.rs
@@ -273,7 +273,7 @@ fn update_atomic_given_key(
                 } => {
                     let key = match key_value {
                         TAtomic::TLiteralString { value } => Some(DictKey::String(value.clone())),
-                        TAtomic::TLiteralInt { value } => Some(DictKey::Int(*value as u32)),
+                        TAtomic::TLiteralInt { value } => Some(DictKey::Int(*value as u64)),
                         TAtomic::TEnumLiteralCase {
                             enum_name,
                             member_name,
